+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.4] - 2024-05-28
+
+### Fixed
+
+- jarPackageDiscovery class loader
+
 ## [1.4.3] - 2024-05-28
 
 ### Changed